requirements.txt
astroid==2.5.2
autopep8==1.5.6
base==1.0.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
cryptography==3.4.7
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
Flask-PyMongo==2.3.0
idna==3.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.20.2
pandas==1.2.3
props==0.0.3
protobuf==3.15.7
pycodestyle==2.7.0
pycparser==2.20
pylint==2.7.4
pymongo==3.11.3
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==2.0.1

Flask-RESTful==0.3.7
google-cloud-firestore==1.6.1
gunicorn==20.0.4
